# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LC Management API",
        "description": "LC Management System API - Complete system for managing contracts, buyers, orders, and shipments. Includes contract management, order processing with cost tracking, shipment tracking with summary statistics, and comprehensive reporting capabilities.",
        "contact": {
            "name": "API Support",
            "email": "support@lcmanagement.example.com"
        },
        "version": "1.2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server (Alternative)"
        }
    ],
    "paths": {
        "/api/b2b-lc": {
            "get": {
                "tags": [
                    "B2B LC"
                ],
                "summary": "List all B2B LCs",
                "description": "Get paginated list of Back-to-Back Letter of Credits with filters",
                "operationId": "ec5103f14f90f87dbf8f1326e23dcff4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pi_number",
                        "in": "query",
                        "description": "Search by PI number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "supplier",
                        "in": "query",
                        "description": "Filter by supplier",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "B2B LC"
                ],
                "summary": "Create new B2B LC",
                "description": "Create a new Back-to-Back Letter of Credit",
                "operationId": "184eaac5be2270b84ede48a54346cb09",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pi_number",
                                    "order_id",
                                    "costing_detail_id",
                                    "issue_date",
                                    "expiry_date"
                                ],
                                "properties": {
                                    "pi_number": {
                                        "type": "string",
                                        "example": "PI-2025-001"
                                    },
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "costing_detail_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "supplier_info": {
                                        "type": "object"
                                    },
                                    "beneficiary_bank_info": {
                                        "type": "object"
                                    },
                                    "pi_value": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "payment_terms": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "draft"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "B2B LC created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/b2b-lc/{id}": {
            "get": {
                "tags": [
                    "B2B LC"
                ],
                "summary": "Get B2B LC details",
                "description": "Get detailed information about a specific B2B LC",
                "operationId": "3b3907d6ffa3cae729b9cb1fb868e66d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "B2B LC not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "B2B LC"
                ],
                "summary": "Update B2B LC",
                "description": "Update an existing B2B LC",
                "operationId": "c94f7c19cdd9edc5c58d50be7005a121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "expiry_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "pi_value": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "supplier_info": {
                                        "type": "object"
                                    },
                                    "payment_terms": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "B2B LC updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "B2B LC not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "B2B LC"
                ],
                "summary": "Delete B2B LC",
                "description": "Remove a B2B LC from the system",
                "operationId": "f0eb2ec1b02dbcc237613e1ae365f2f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "B2B LC deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "B2B LC not found"
                    }
                }
            }
        },
...
    },
...
}